move EnsemblData to objects.py
