+++ src/SOPRANO/utils/url_utils.py
@@ -3,7 +3,6 @@ import pathlib
 import pickle as pk
 import shutil
 from pathlib import Path
-from typing import Set
 
 import requests
 from clint.textui import progress
@@ -212,187 +211,3 @@ def build_ensembl_urls(species: str, reference: str):
     primary_assembly_url = f"{_url}.primary_assembly.fa.gz"
 
     return {"toplevel": toplevel_url, "primary_assembly": primary_assembly_url}
-
-
-class _GatherReferences:
-    # Urls
-    toplevel_url: str
-    primary_assembly_url: str
-
-    # Params
-    species: str
-    reference: str
-
-    # Status
-    toplevel_gz_done = Set[int]
-    toplevel_fa_done = Set[int]
-    toplevel_fai_done = Set[int]
-    primary_assembly_gz_done = Set[int]
-    primary_assembly_fa_done = Set[int]
-    primary_assembly_fai_done = Set[int]
-    sizes_done = Set[int]
-
-    def _dest_directory(self, release: int):
-        return Directories.data(self.species) / f"{release}_{self.reference}"
-
-    def _dest_fa_gz(self, release: int, _toplevel: bool):
-        return self._dest_directory(release) / filename_from_url(
-            self.toplevel_url if _toplevel else self.primary_assembly_url
-        ).format(RELEASE=release)
-
-    def _dest_fa(self, release: int, _toplevel: bool):
-        return self._dest_fa_gz(release, _toplevel).with_suffix("")
-
-    def _dest_fai(self, release: int, _toplevel: bool):
-        return self._dest_fa_gz(release, _toplevel).with_suffix(".fai")
-
-    def dest_chrom(self, release: int, _toplevel: bool):
-        return self._dest_fa(release, _toplevel).with_suffix(".chrom")
-
-    def toplevel_fa_gz_path(self, release: int):
-        return self._dest_fa_gz(release, _toplevel=True)
-
-    def toplevel_fa_path(self, release: int):
-        return self._dest_fa(release, _toplevel=True)
-
-    def toplevel_fai_path(self, release: int):
-        return self._dest_fai(release, _toplevel=True)
-
-    def toplevel_chrom_path(self, release: int):
-        return self.toplevel_fa_path(release).with_suffix(".chrom")
-
-    def primary_assembly_fa_gz_path(self, release: int):
-        return self._dest_fa_gz(release, _toplevel=False)
-
-    def primary_assembly_fa_path(self, release: int):
-        return self._dest_fa(release, _toplevel=False)
-
-    def primary_assembly_fai_path(self, release: int):
-        return self._dest_fai(release, _toplevel=False)
-
-    def _download(self, release: int, _toplevel):
-        if _toplevel:
-            source_url = self.toplevel_url
-            dest_path = self.toplevel_fa_gz_path(release)
-            decompressed_path = self.toplevel_fa_path(release)
-        else:
-            source_url = self.primary_assembly_url
-            dest_path = self.primary_assembly_fa_gz_path(release)
-            decompressed_path = self.toplevel_fa_path(release)
-
-        if not (decompressed_path.exists() or dest_path.exists()):
-            dest_path.parent.mkdir(parents=True, exist_ok=True)
-            check_ensembl_file_url(source_url, release)
-            download_from_url(
-                source_url.format(RELEASE=release),
-                target_path=dest_path,
-            )
-
-    def _check_release_ok(self, release):
-        min_release = find_earliest_release(self.toplevel_url)
-        max_release = find_latest_release(self.toplevel_url)
-
-        if not (min_release <= release <= max_release):
-            raise ValueError(release)
-
-    def download_toplevel(self, release):
-        if release not in self.toplevel_gz_done:
-            self._check_release_ok(release)
-
-            if not self.toplevel_fa_gz_path(release).exists():
-                self._download(release, _toplevel=True)
-
-            self.toplevel_gz_done.add(release)
-
-    def download_primary_assembly(self, release):
-        if release not in self.primary_assembly_gz_done:
-            self._check_release_ok(release)
-
-            if self.primary_assembly_fa_gz_path(release).exists():
-                self._download(release, _toplevel=False)
-
-            self.primary_assembly_gz_done.add(release)
-
-    def decompress_toplevel(self, release):
-        if release not in self.toplevel_fa_done:
-            if not self.toplevel_fa_path(release).exists():
-                decompress(self.toplevel_fa_gz_path(release))
-
-            self.toplevel_fa_done.add(release)
-
-    def decompress_primary_assembly(self, release):
-        if release not in self.primary_assembly_fa_done:
-            if not self.primary_assembly_fa_path(release).exists():
-                decompress(self.primary_assembly_fa_gz_path(release))
-
-            self.primary_assembly_fa_done.add(release)
-
-    def compute_chrom_sizes(self, release):
-        if release not in self.sizes_done:
-            if not self.toplevel_chrom_path(release).exists():
-                compute_chrom_sizes(self.toplevel_fai_path(release))
-
-            self.sizes_done.add(release)
-
-    def compute_fasta_index_toplevel(self, release):
-        if release not in self.toplevel_fai_done:
-            if not self.toplevel_fai_path(release).exists():
-                compute_fasta_index(self.toplevel_fa_path(release))
-
-            self.toplevel_fai_done.add(release)
-
-    def compute_fasta_index_primary_assembly(self, release):
-        if release not in self.primary_assembly_fai_done:
-            if not self.primary_assembly_fai_path(release).exists():
-                compute_fasta_index(self.primary_assembly_fa_path(release))
-
-            self.primary_assembly_fai_done.add(release)
-
-    def compute_all_toplevel(self, release):
-        self.download_toplevel(release)
-        self.decompress_toplevel(release)
-        self.compute_fasta_index_toplevel(release)
-        self.compute_chrom_sizes(release)
-
-    def compute_all_primary_assembly(self, release):
-        self.download_primary_assembly(release)
-        self.decompress_primary_assembly(release)
-        self.compute_fasta_index_primary_assembly(release)
-
-
-class EnsemblData(_GatherReferences):
-    def __init__(self, species: str, reference: str, _init_urls=True):
-        self.species = species
-        self.reference = reference
-
-        if _init_urls:
-            url_dict = build_ensembl_urls(species, reference)
-            self.toplevel_url = url_dict["toplevel"]
-            self.primary_assembly_url = url_dict["primary_assembly"]
-
-    @classmethod
-    def homo_sapiens_GRCh38(cls):
-        return cls("homo_sapiens", "GRCh38")
-
-    @classmethod
-    def homo_sapiens_GRCh37(cls):
-        # GRCh37 is actually has a deviant url structure, so manually set here
-        toplevel_url = (
-            "https://ftp.ensembl.org/pub/grch37/release-{RELEASE}/"
-            "fasta/homo_sapiens/dna/"
-            "Homo_sapiens.GRCh37.dna.toplevel.fa.gz"
-        )
-
-        primary_assembly_url = (
-            "https://ftp.ensembl.org/pub/grch37/release-{RELEASE}/"
-            "fasta/homo_sapiens/dna/"
-            "Homo_sapiens.GRCh37.dna.primary_assembly.fa.gz"
-        )
-
-        species = "homo_sapiens"
-        reference = "GRCh37"
-
-        obj = cls(species, reference, _init_urls=False)
-        obj.toplevel_url = toplevel_url
-        obj.primary_assembly_url = primary_assembly_url
-        return obj
